backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
redis>=5.0.1
sentence-transformers>=2.2.2
numpy>=1.24.0
pydantic>=2.4.2
pinecone-client>=3.0.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.25.0